bootcamp/requirements.txt
jupyterlab<=4.2.4
langchain==0.2.10
langchain-openai==0.1.17
python-dotenv == 1.0.1
langchain-groq == 0.1.6
langchain-community == 0.2.10
faiss-cpu == 1.8.0
pypdf==5.0.1
langchain-chroma ==0.1.2